backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
facedb
django-cors-headers
pytz